add momo payment

+++ clinic/src/main/java/com/hb/controllers/LabTestController.java
@@ -4,7 +4,6 @@
  */
 package com.hb.controllers;
 
-import com.hb.pojo.LabTests;
 import com.hb.service.LabTestService;
 import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,7 +12,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -45,7 +43,7 @@ public class LabTestController {
 
     @DeleteMapping("/lab-tests/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable(value = "id") Integer id) {
+    public void delete(@PathVariable(value = "id") Long id) {
         labTestService.deleteLabTest(id);
     }
 }
